Replace old genReadme
